requirements.txt
einops==0.6.1
h5py==3.9.0
matplotlib==3.7.2
scipy==1.11.1
tensorboard==2.13.0
# torch==2.0.1
# torchaudio==2.0.2
# torchvision==0.15.2
torchmetrics==1.0.1
tqdm==4.65.0
yacs==0.1.8
